Remove legacy Cosmos persistence and V2 migration tooling.
Delete RedditPodcastPoster.Persistence.Legacy, LegacyPodcastToV2Migration, and dual-path console apps; consolidate CosmosDb tools on current persistence only.
