+++ .github/copilot-instructions.md
@@ -6,9 +6,8 @@
 - User does not want additional Episode members for compact IDs; use existing Spotify/YouTube/Apple IDs and derive Apple episode slug from Apple URL via regex.
 - In migration/infrastructure docs, use `LookUps` as the Cosmos container name instead of `knownTerms`, and store the single `KnownTerms` typed item in `LookUps`.
 - EliminationTerms should be stored in the `LookUps` container, and infrastructure should include a dedicated `PushSubscriptions` container.
-- Prefer using `PodcastEpisodeV2` and V2 models across the codebase, except in `PodcastRepository` and `LegacyPodcastToV2Migration`.
-- Prefer migrating code to V2 podcast/episode models everywhere possible; keep legacy models only in `PodcastRepository` and `LegacyPodcastToV2Migration`.
-- Use V2 repository implementations when both legacy and V2 repositories exist in this codebase.
+- Prefer using `PodcastEpisodeV2` and V2 models across the codebase.
+- Use the canonical split-container repository implementations in `RedditPodcastPoster.Persistence`.
 - The project is not using soft-delete for Azure Search indexing.
 - Episode ID is globally unique in this project.
 - Documents under `docs/migration` must be kept up to date when making related changes.
@@ -25,9 +24,6 @@
 
 ## Azure Search Indexing
 - When automating Azure Search indexer reruns, wait for the newly triggered run to start (using status start-time correlation) before evaluating completion to avoid premature retriggers and 409 conflicts.
-
-## Migration Verification
-- When resetting the V2 database for migration verification, use the built-in teardown flags of the `LegacyPodcastToV2Migration` app rather than manual deletes.
 
 ## Debugging Preferences
 - User prefers evidence and ongoing debugging notes to be added to `debug-session-summary.md` instead of being expanded in chat.
