--- Class-Libraries/RedditPodcastPoster.Persistence.Legacy/CosmosDbClientFactory.cs
@@ -1,32 +1,0 @@
-using Microsoft.Azure.Cosmos;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using RedditPodcastPoster.Persistence.Abstractions;
-
-namespace RedditPodcastPoster.Persistence.Legacy;
-
-public class CosmosDbClientFactory(
-    IJsonSerializerOptionsProvider jsonSerializerOptionsProvider,
-    IOptions<CosmosDbSettings> settings,
-#pragma warning disable CS9113 // Parameter is unread.
-    ILogger<CosmosDbClientFactory> logger)
-#pragma warning restore CS9113 // Parameter is unread.
-    : ICosmosDbClientFactory
-{
-    private readonly CosmosDbSettings _settings = settings.Value;
-
-    public CosmosClient Create()
-    {
-        var cosmosClientOptions = new CosmosClientOptions
-        {
-            UseSystemTextJsonSerializerWithOptions = jsonSerializerOptionsProvider.GetJsonSerializerOptions()
-        };
-        if (_settings.UseGateway.HasValue && _settings.UseGateway.Value)
-        {
-            cosmosClientOptions.ConnectionMode = ConnectionMode.Gateway;
-        }
-
-        CosmosClient client = new(_settings.Endpoint, _settings.AuthKeyOrResourceToken, cosmosClientOptions);
-        return client;
-    }
-}
