--- Class-Libraries/RedditPodcastPoster.Persistence.Legacy/CosmosDbExtensions.cs
@@ -1,23 +1,0 @@
-using Microsoft.Azure.Cosmos;
-
-namespace RedditPodcastPoster.Persistence.Legacy;
-
-internal static class CosmosDbExtensions
-{
-    /// <summary>
-    /// Convert a feed iterator to IAsyncEnumerable
-    /// </summary>
-    /// <typeparam name="TModel"></typeparam>
-    /// <param name="setIterator"></param>
-    /// <returns></returns>
-    public static async IAsyncEnumerable<TModel> ToAsyncEnumerable<TModel>(this FeedIterator<TModel> setIterator)
-    {
-        while (setIterator.HasMoreResults)
-        {
-            foreach (var item in await setIterator.ReadNextAsync())
-            {
-                yield return item;
-            }
-        }
-    }
-}
