--- Class-Libraries/RedditPodcastPoster.Persistence.Legacy/FileKeyWrapper.cs
@@ -1,9 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace RedditPodcastPoster.Persistence.Legacy;
-
-public class FileKeyWrapper
-{
-    [JsonPropertyName("fileKey")]
-    public required string FileKey { get; set; }
-}
