--- Class-Libraries/RedditPodcastPoster.Persistence.Legacy/ICosmosDbClientFactory.cs
@@ -1,8 +1,0 @@
-using Microsoft.Azure.Cosmos;
-
-namespace RedditPodcastPoster.Persistence.Legacy;
-
-public interface ICosmosDbClientFactory
-{
-    CosmosClient Create();
-}
