--- Class-Libraries/RedditPodcastPoster.Persistence.Legacy/ICosmosDbRepository.cs
@@ -1,7 +1,0 @@
-using RedditPodcastPoster.Persistence.Abstractions;
-
-namespace RedditPodcastPoster.Persistence.Legacy;
-
-public interface ICosmosDbRepository : IDataRepository
-{
-}
