--- Class-Libraries/RedditPodcastPoster.Persistence.Legacy/IEpisodeMatcher.cs
@@ -1,8 +1,0 @@
-using System.Text.RegularExpressions;
-
-namespace RedditPodcastPoster.Persistence.Legacy;
-
-public interface IEpisodeMatcher
-{
-    bool IsMatch(Episode existingEpisode, Episode episodeToMerge, Regex? episodeMatchRegex);
-}
