--- Class-Libraries/RedditPodcastPoster.Persistence.Legacy/IPushSubscriptionRepository.cs
@@ -1,10 +1,0 @@
-using RedditPodcastPoster.Models;
-
-namespace RedditPodcastPoster.Persistence.Legacy;
-
-public interface IPushSubscriptionRepository
-{
-    Task Save(PushSubscription pushSubscription);
-    IAsyncEnumerable<PushSubscription> GetAll();
-    Task Delete(PushSubscription pushSubscription);
-}
