+++ Console-Apps/JsonSplitCosmosDbUploader/JsonSplitCosmosDbUploadProcessor.cs
@@ -5,7 +5,7 @@ using RedditPodcastPoster.JsonSplitCosmosDbUploader;
 using RedditPodcastPoster.Persistence.Abstractions;
 using Episode = RedditPodcastPoster.Models.Episode;
 using IPodcastRepository = RedditPodcastPoster.Persistence.Abstractions.IPodcastRepository;
-using Podcast = RedditPodcastPoster.Persistence.Legacy.Podcast;
+using EmbeddedPodcast = JsonSplitCosmosDbUploader.EmbeddedFileFormat.EmbeddedPodcast;
 
 namespace JsonSplitCosmosDbUploader;
 
@@ -19,7 +19,7 @@ public class JsonSplitCosmosDbUploadProcessor(
 {
     public async Task Run(JsonSplitCosmosDbUploadRequest request)
     {
-        var sourcePodcast = await fileRepository.Read<Podcast>(Path.GetFileNameWithoutExtension(request.FileName));
+        var sourcePodcast = await fileRepository.Read<EmbeddedPodcast>(Path.GetFileNameWithoutExtension(request.FileName));
         if (sourcePodcast != null)
         {
             logger.LogInformation("'{EpisodesCount}' episodes.", sourcePodcast.Episodes.Count);
