--- Console-Apps/LegacyPodcastToV2Migration/Properties/launchSettings.json
@@ -1,8 +1,0 @@
-{
-  "profiles": {
-    "LegacyPodcastToV2Migration": {
-      "commandName": "Project",
-      "commandLineArgs": "--teardown true"
-    }
-  }
-}
