+++ RedditPodcastPoster.slnx
@@ -35,8 +35,6 @@
     <Project Path="Class-Libraries/RedditPodcastPoster.Models/RedditPodcastPoster.Models.csproj" />
 
     <Project Path="Class-Libraries/RedditPodcastPoster.Persistence.Abstractions/RedditPodcastPoster.Persistence.Abstractions.csproj" />
-    <Project Path="Class-Libraries/RedditPodcastPoster.Persistence.Legacy/RedditPodcastPoster.Persistence.Legacy.csproj" />
-
     <Project Path="Class-Libraries/RedditPodcastPoster.Persistence/RedditPodcastPoster.Persistence.csproj" />
 
     <Project Path="Class-Libraries/RedditPodcastPoster.PodcastServices.Abstractions/RedditPodcastPoster.PodcastServices.Abstractions.csproj" />
@@ -173,10 +171,6 @@
     <Project Path="Console-Apps/LanguagesPublisher/LanguagesPublisher.csproj" />
 
     <Project Path="Console-Apps/LaunchSettingsToAppSettings/LaunchSettingsToAppSettings.csproj" />
-    <Project Path="Console-Apps/LegacyPodcastToV2Migration/LegacyPodcastToV2Migration.csproj">
-      <Build Solution="Debug|*" Project="false" />
-    </Project>
-
     <Project Path="Console-Apps/MachineAuth0/MachineAuth0.csproj" />
 
     <Project Path="Console-Apps/Poster/Poster.csproj" />
